Modified the Converter to ignore time-only trackpoints that are the first trackpoint in a track.
I was emailed a workout that was one massive (22km) with lots of <Track> elements.  The time difference
between the final <Trackpoint> of a <Track> and the first <Trackpoint> of the next <Track> often indicated
that there was a pause.  However - the first <Trackpoint> of each <Track> was just a time-only
<Trackpoint> so we were generating a pause using that as the start-time of the pause which meant the proper
pause was missed.

We now ignore time-only <Trackpoint> if it's the first <Trackpoint> in the <Track>.  The potential-pause
will pick up the proper pause duration if one is required.