Further UI improvements (cookies, xml download).
Added a "Remember Details" checkbox which when checked and the
after a successfull convert & upload will save much of the form
state using cookies.  When the user arrives at the page next time
the cookies will be looked at and used to populate the form fields.

Also reintroduced the nike+ xml download option.  To download the
nike+ xml the user needs to collapse both the simple and advanced
authentication fieldsets.