Fixed bug in ConvertTcx where we were using an old reference to previousTp.
In ConvertTcx, when validating the lap data, in an iteration of
trackpoints which strips zero-distance and slowest-pace was referring
to a variable previousTp, which had not been updated since a previous
iteration.  It was a simple fix once I realised what an idiot I'd been.

I also have removed support for adding actual pause/resumes as garmin
seem to be inconsistent with marking these properly in the tcx file.
I know just remove null/zero-distance trackpoints and decrement further
trackpoint durations silently and only if required by to ensure lap
validity.