We now ignore trackponts where their predecessor had <= distance.
We now only add trackpoints where the distance is greater than the most
recent distance (other values, such as a load of 0.00m at the start of
a workout indicate no satellites (I think).

We discard the trackpoint, as otherwise it'll be used in the lap
validation as a 'slowest pace' trackpoint when it shouldn't be.