Add NikePlus.syncData() for "run" JSON and "gpxXML" XML data
The Nike+ iPhone application sends a multipart/form-data HTTP POST with two
parts:
1) GPX (http://www.topografix.com/GPX/1/1) representation of activity ("run").
2) Additional XML data for the nike+ website ("gpxXML").

I've added a new NikePlus.syncData(...) override which takes a gson
JsonElement and a xjc-generated GpxType.  I've marked the previous version
(which is still used) as @Deprecated and added unit tests for the new
version.  There is also some re-factoring of other classes to make this
tidier.